增加IO功能，完善代码结构
