+++ IOManager.cpp
@@ -1,0 +1,7 @@
+#include "IOManager.h"
+
+bool IOManager::eof()
+{
+	ifstream ifs(file_name);
+	return ifs.eof();
+}
