+++ Storage.cpp
@@ -1,0 +1,2 @@
+#include "Storage.h"
+const char* Storage::FILENAME = "data.txt";
