+++ user.cpp
@@ -1,0 +1,2 @@
+#include "user.h"
+vector<user*> userVector;
